trocando o tipo de int_t para os IDs

+++ src/artesao.cpp
@@ -30,7 +30,7 @@ Artesao::Artesao(){
     this->start_ociosity = 0;
 }
 /* SET METHODS */
-void Artesao::set_id(u_int8_t id){
+void Artesao::set_id(u_int16_t id){
     this->id = id;
 }
 void Artesao::set_situation(state_art situation){
@@ -43,7 +43,7 @@ void Artesao::set_start_ociosity(u_int32_t start_ociosity){
     this->start_ociosity = start_ociosity;
 }
 /* GET METHODS */
-u_int8_t  Artesao::get_id(){
+u_int16_t  Artesao::get_id(){
     return this->id;
 }
 state_art Artesao::get_situation(){
