+++ src/include/especialista.h
@@ -21,18 +21,18 @@ class Especialista{
     static Especialista* get_free();
     void                 add_esp();
     /* SET METHODS */
-    void      set_id(u_int8_t id);
+    void      set_id(u_int16_t id);
     void      set_situation(state_esp situation);
     void      set_time_ociosity (u_int32_t time_ociosity);
     void      set_start_ociosity(u_int32_t start_ociosity);
     /* GET METHODS */
-    u_int8_t  get_id();
+    u_int16_t get_id();
     state_esp get_situation();
     u_int32_t get_time_ociosity ();
     u_int32_t get_start_ociosity();
 
   private:
-    u_int8_t  id;
+    u_int16_t id;
     state_esp situation;
     u_int32_t time_ociosity;
     u_int32_t start_ociosity;
